+++ model/FEM/Boundaries/ExternalBoundary.h
@@ -36,7 +36,7 @@ namespace model
                 if(eIter->second.isBoundaryElement())
                 {
                     const std::vector<int> boundaryFaces=eIter->second.boundaryFaces();
-                    for (int f=0;f<boundaryFaces.size();++f)
+                    for (size_t f=0;f<boundaryFaces.size();++f)
                     {
                         temp.emplace_back(&eIter->second,boundaryFaces[f]);
                     }
