+++ model/FEM/WeakForms/BilinearWeakForm.h
@@ -98,7 +98,7 @@ namespace model
             
             
             const auto t0= std::chrono::system_clock::now();
-            for (int k=0;k<domain.size();++k)
+            for (size_t k=0;k<domain.size();++k)
             {
                 const ElementType& ele(domain.element(k));
                 ElementMatrixType ke(ElementMatrixType::Zero());
