+++ model/Threads/EqualIteratorRange.h
@@ -27,7 +27,7 @@ namespace model
         typedef typename ContainerType::iterator IterType;
         
         /**********************************************************************/
-        EqualIteratorRange(const IterType& first, const IterType& last, const int& nRanges)
+        EqualIteratorRange(const IterType& first, const IterType& last, const size_t& nRanges)
         {
             assert(nRanges>0 && "NUMBER OF RANGES MUST BE >0");
             
@@ -42,7 +42,7 @@ namespace model
             IterType currentBegin(first);
             IterType currentEnd(first);
             
-            for (int i=0;i<nRanges;++i)
+            for (size_t i=0;i<nRanges;++i)
             {
                 std::advance(currentEnd,(i<remainder)? quotient+1 : quotient);
                 this->emplace_back(currentBegin,currentEnd);
@@ -62,7 +62,7 @@ namespace model
         typedef typename ContainerType::const_iterator IterType;
         
         /**********************************************************************/
-        EqualConstIteratorRange(const IterType& first, const IterType& last, const int& nRanges)
+        EqualConstIteratorRange(const IterType& first, const IterType& last, const size_t& nRanges)
         {
             assert(nRanges>0 && "NUMBER OF RANGES MUST BE >0");
             
@@ -77,7 +77,7 @@ namespace model
             IterType currentBegin(first);
             IterType currentEnd(first);
             
-            for (int i=0;i<nRanges;++i)
+            for (size_t i=0;i<nRanges;++i)
             {
                 std::advance(currentEnd,(i<remainder)? quotient+1 : quotient);
                 this->emplace_back(currentBegin,currentEnd);
